requirements.txt
aniso8601==7.0.0
asgiref==3.2.7
certifi==2020.4.5.1
Django==3.0.7
django-environ==0.4.5
djangorestframework==3.11.0
graphene==2.1.8
graphene-django==2.10.1
graphql-core==2.3.2
graphql-relay==2.0.1
promise==2.3
psycopg2==2.8.5
pytz==2020.1
Rx==1.6.1
singledispatch==3.4.0.3
six==1.15.0
sqlparse==0.3.1
